requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==4.2.7
django-cors-headers==4.4.0
django-cron==0.6.0
django-filter==24.3
django-modeltranslation==0.19.9
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.7
msgpack==1.1.0
phonenumbers==8.13.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
redis==5.1.1
service-identity==24.1.0
setuptools==75.1.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.1.0
